Bug fixes and more deep learning preparation.
Abstracted room to be a general class, suitable for use in both web and
deep learning.

Fixed a bug with the "blocked tile" array being global instead of
game-specific.

Created a function to build an emittable object ready to be pushed into
a neural network.

Added terrain functions necessary for said emittable object.

Added inspectable handling for the new map types that exist in game.

Made a file for the deep learning interface and set a skeleton for it.

Fixed a logic bug in the RuleBasedAi regarding relative_tick.